+++ src/ansys/aedt/core/extensions/common/advanced_fields_calculator.py
