+++ src/ansys/aedt/core/extensions/common/configure_layout.py
@@ -27,7 +27,7 @@ import tkinter
 
 from ansys.aedt.core.extensions.misc import ExtensionCommon
 from ansys.aedt.core.extensions.misc import get_arguments
-from ansys.aedt.core.extensions.project.resources.configure_layout.master_ui import ConfigureLayoutExtension
+from ansys.aedt.core.extensions.common.resources.configure_layout.master_ui import ConfigureLayoutExtension
 
 if __name__ == "__main__":  # pragma: no cover
     args = get_arguments(ConfigureLayoutExtension.EXTENSION_TITLE)
