+++ src/ansys/aedt/core/extensions/common/images/large/version_manager.png
