+++ src/ansys/aedt/core/extensions/common/images/large/via_design_icon.png
