+++ src/ansys/aedt/core/extensions/common/resources/configure_layout/data_class.py
