+++ src/ansys/aedt/core/extensions/common/resources/configure_layout/tab_main.py
