+++ src/ansys/aedt/core/extensions/common/resources/configure_layout/template.py
